+++ Sources/App/Models/Basket/RemoveProduct/RemoveProductResponse.swift
